ocr_v2/requirements.txt
PyMuPDF>=1.23.0
opencv-python>=4.8.0
numpy>=1.24.0
easyocr>=1.7.1
google-generativeai>=0.4.0
python-dotenv>=1.0.0
